api/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
click==8.1.7
colorama==0.4.6
exceptiongroup @ file:///C:/b/abs_c5h1o1_b5b/croot/exceptiongroup_1706031441653/work
fastapi==0.109.0
fastapi-sessions==0.3.2
greenlet==3.0.3
h11==0.14.0
idna==3.6
itsdangerous==2.2.0
mysql-connector-python==8.3.0
pydantic==2.5.3
pydantic_core==2.14.6
PyMySQL==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.35.1
typing_extensions==4.11.0
uvicorn==0.29.0
